[networking/ufw/firewall_tables] Split out firewall from networking
Drop fallback to nat as that can bring in more kernel modules
Left ip/6-table save in networking for now as it can be filtered by
networking options and I don't want to break that.

Resolves: #2348
